\name{HT29R.expNames}
\alias{HT29R.expNames}
\docType{data}
\title{
Benchmark Experiment Names
}
\description{
Labels of individual HT-29 screen experiments
}
\usage{data("HT29R.expNames")}
\format{
  A vector of 6 strings, containing each the name of one experiment.
}
\examples{
data(HT29R.expNames)
print(HT29R.expNames)
}
\keyword{datasets}
